+++ src/application/VcsSplashScreen.vue
@@ -28,36 +28,34 @@
       </v-card-text>
 
       <v-card-actions>
-        <v-row no-gutters>
-          <v-col class="text-right">
-            <VcsFormButton
-              color="primary"
-              variant="filled"
-              @click="exitScreen"
-              :disabled="options.acceptInput && !checkBox"
-              >{{
-                $st(
-                  options.buttonTitle || 'components.splashScreen.buttonTitle',
-                )
-              }}</VcsFormButton
-            >
-          </v-col>
-        </v-row>
+        <div class="d-flex gc-2 w-100 justify-end">
+          <VcsFormButton
+            v-if="
+              options.secondaryButtonTitle && options.secondaryCallbackOptions
+            "
+            @click="secondaryButtonClicked"
+          >
+            {{ $st(options.secondaryButtonTitle) }}
+          </VcsFormButton>
+          <VcsFormButton
+            color="primary"
+            variant="filled"
+            @click="exitScreen"
+            :disabled="options.acceptInput && !checkBox"
+            >{{
+              $st(options.buttonTitle || 'components.splashScreen.buttonTitle')
+            }}</VcsFormButton
+          >
+        </div>
       </v-card-actions>
     </v-card>
   </v-dialog>
 </template>
 
 <script>
-  import {
-    VDialog,
-    VCard,
-    VCardText,
-    VCardActions,
-    VRow,
-    VCol,
-  } from 'vuetify/components';
-  import { computed, ref } from 'vue';
+  import { VDialog, VCard, VCardText, VCardActions } from 'vuetify/components';
+  import { computed, ref, inject } from 'vue';
+  import { executeCallbacks } from '../callback/vcsCallback.js';
   import VcsFormButton from '../components/buttons/VcsFormButton.vue';
   import VcsCheckbox from '../components/form-inputs-controls/VcsCheckbox.vue';
   import VcsMarkdown from '../components/form-output/VcsMarkdown.vue';
@@ -72,8 +70,6 @@
       VCardActions,
       VcsFormButton,
       VcsCheckbox,
-      VRow,
-      VCol,
       VcsMarkdown,
     },
     props: {
@@ -87,13 +83,23 @@
       },
     },
     setup(props, { emit }) {
+      const app = inject('vcsApp');
       const localValue = useProxiedAtomicModel(props, 'modelValue', emit);
 
       const checkBox = ref(false);
-      const exitScreen = () => {
+      function exitScreen() {
         localValue.value = false;
         checkBox.value = false;
-      };
+        if (Array.isArray(props.options.exitCallbackOptions)) {
+          executeCallbacks(app, props.options.exitCallbackOptions);
+        }
+      }
+
+      function secondaryButtonClicked() {
+        if (Array.isArray(props.options.secondaryCallbackOptions)) {
+          executeCallbacks(app, props.options.secondaryCallbackOptions);
+        }
+      }
 
       const position = computed(() => ({
         width: props.options.position?.width || 800,
@@ -105,6 +111,7 @@
       return {
         localValue,
         exitScreen,
+        secondaryButtonClicked,
         checkBox,
         position,
       };