+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 - Suggestions implemented for search interfaces that provide string suggestions.
 - Fixes styling issues with IframeComponent, BalloonComponent and WindowComponent
 - Fixes issue with VListItem subtitle opacity by adding `list-item-subtitle-opacity` to vuetify variables
+- Remove Config pattern check for modules in `init.js`
+- Add two more `VcsCallback` extensions: `AddModuleCallback` and `RemoveModuleCallback`
+- Extend `splashScreen` configuration with in `uiConfig`:
+  - added option `exitCallbackOptions` to configure an array of callback options executed on click a new secondary button
+  - added option `secondaryCallbackOptions` to configure an array of callback options executed on click of the primary (exit) button
+  - added option `secondaryButtonTitle` to configure the new secondary button title
 
 # 6.0.11